// src/domain/recetaMedica/types.ts
export interface RecetaMedicaDocument {
  numero: string
  fechaDia: string
  fechaMes: string
  fechaAnio: string
  nombresApellidos: string
  edad: string
  ci: string
  hc: string
  diagnostico: string
  cie10: string
  antAlergicos: string
  rp: string
  indicaciones: string
  nombreMedico: string
  celularMedico: string
  regMedico: string
}